# PERF: Add index for chat unread counts query